frontend for question page
