+++ packages/client/src/containers/QuestionPage/CheckinQuestions.css
@@ -1,0 +1,187 @@
+:root {
+  --primary-color: #1d4ed8;
+  --background-color: #f7f7f7;
+  --primary-background-color: #fff;
+  --text-color: #fffff;
+  --light-subheading-text-color: #64748b;
+  --primary-disabled-text: #c2c0ff;
+  --light-sub-text-color: #a0abbb;
+}
+
+* {
+  padding: 0;
+  margin: 0;
+  box-sizing: border-box;
+}
+
+body {
+  height: 100vh;
+  background-color: var(--background-color);
+}
+
+.dashboard-container {
+  display: flex;
+}
+
+.left-nav-bar-container {
+  display: flex;
+  padding: 40px 0px;
+  flex-direction: column;
+  align-items: center;
+  gap: 80px;
+  flex-shrink: 0;
+  border-radius: 0px 40px 40px 0px;
+  background: var(--primary-color);
+}
+
+.left-nav-bar {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  align-items: flex-start;
+  flex-shrink: 0;
+  flex-grow: 1;
+}
+
+.left-nav-bar-top {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+}
+
+.left-nav-bar-top .menu-icon:nth-child(1) {
+  color: white;
+}
+
+.menu-icon {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  gap: 18.204px;
+  flex: 1 0 0;
+  color: var(--fade-color);
+}
+
+.dashboard {
+  color: var(--text-color);
+}
+
+.fade {
+  color: var(--fade-color);
+}
+
+.left-nav-bar-bottom {
+  display: flex;
+  align-items: center;
+}
+
+.content-container {
+  display: flex;
+  flex-direction: column;
+  align-items: stretch;
+  gap: 80px;
+  flex-shrink: 0;
+  flex-grow: 1;
+  height: 100vh;
+  overflow-x: hidden;
+}
+
+.top-nav-bar {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  align-items: center;
+  align-self: stretch;
+  padding: 40px 80px;
+  background: var(--primary-background-color);
+}
+
+.heading-text {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  gap: 12px;
+  flex: 1 0 0;
+}
+
+.sub-heading {
+  color: var(--light-subheading-text-color);
+}
+
+.subtext {
+  color: var(--light-sub-text-color);
+}
+
+.image-container {
+  display: flex;
+  align-items: center;
+  gap: 12px;
+}
+
+.placeholder-image {
+  width: 50px;
+  height: 50px;
+  flex-shrink: 0;
+  border-radius: 50px;
+  border: 3px solid #1d4ed8;
+  background: linear-gradient(
+      0deg,
+      rgba(217, 217, 217, 0.1) 0%,
+      rgba(217, 217, 217, 0.1) 100%
+    ),
+    lightgray 50% / cover no-repeat;
+}
+
+.name-mail-container {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  flex: 1 0 0;
+}
+
+.question-list-container {
+  display: flex;
+  flex-direction: column;
+  flex: 1 0 0;
+  margin-left: 30px;
+  margin-right: 30px;
+}
+
+.question-list-header {
+  display: flex;
+  align-items: center;
+  justify-content: flex-start;
+  margin-bottom: 20px;
+}
+
+.question-list {
+  list-style-type: none; /* Remove default list styles */
+  background: var(--primary-background-color);
+  padding: 10px 20px;
+}
+
+.question-item {
+  padding: 10px 0; /* Add padding top and bottom */
+}
+
+.question-item:not(:last-child) {
+  border-bottom: 1px solid #ccc; /* Add bottom border to all but the last item */
+}
+
+.question-content {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+}
+
+.question-content .question-actions {
+  display: flex;
+  gap: 10px;
+}
+
+.button-container {
+  display: flex;
+  justify-content: center;
+  margin-top: 20px;
+  margin-bottom: 20px;
+}
